changed locations of generateCard function
